requirements.txt
# MySQL Migration Dependencies

# Database connectivity
pymysql==1.1.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0

# Data processing
polars==0.19.19
pandas==2.1.4

# Configuration
pyyaml==6.0.1

# Logging and utilities
python-dateutil==2.8.2
